pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.0"
description = ""
authors = ["Deborah Foroni <deborah.foroni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.40.2"
pydantic = "^2.10.3"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.48"
mkdocstrings = "^0.27.0"
mkdocstrings-python = "^1.12.2"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
